I think it all works!! Fast Download is very nice too.

+++ maskrcnn_modanet/download.sh
@@ -11,50 +11,62 @@ echo "saving your path location"
 PATH1=$1
 echo $PATH1
 
+FAST=$2
+echo "fast download:"
+echo $FAST
+
 cd $PATH1
 
 mkdir datasets
 cd datasets
 
-
-# download images dataset
-git clone https://github.com/kyamagu/paperdoll
-pwd
-cd paperdoll/data/chictopia
-pwd
-
-if [ ! -d "./photos.lmdb" ]
+if [ ! $FAST ]
 then
-	echo "If you already have the 40GB file lying around, you can stop the download by closing this program now,"
-	echo "putting the photos.lmdb file into ./datasets/paperdoll/data/chictopia"
-	echo "and then restarting this program again so that it thinks it's already downloaded (did you?)"
-	echo "or you could just wait a few hours of your precious time here.."
-	wget -c http://vision.is.tohoku.ac.jp/chictopia2/photos.lmdb.tar
-	tar xf photos.lmdb.tar
-	if [ -d "./photos.lmdb" ]
+	
+	# download images dataset
+	git clone https://github.com/kyamagu/paperdoll
+	pwd
+	cd paperdoll/data/chictopia
+	pwd
+
+	if [ ! -d "./photos.lmdb" ]
 	then
-		rm photos.lmdb.tar
+		echo "If you already have the 40GB file lying around, you can stop the download by closing this program now,"
+		echo "putting the photos.lmdb file into ./datasets/paperdoll/data/chictopia"
+		echo "and then restarting this program again so that it thinks it's already downloaded (did you?)"
+		echo "or you could just wait a few hours of your precious time here.."
+		wget -c http://vision.is.tohoku.ac.jp/chictopia2/photos.lmdb.tar
+		tar xf photos.lmdb.tar
+		if [ -d "./photos.lmdb" ]
+		then
+			rm photos.lmdb.tar
+		else
+			echo "install tar and run again this command!"
+			exit 1
+		fi
+	else echo "photos database already downloaded!"
+	fi
+
+	echo "unzipping database.."
+	gunzip -c chictopia.sql.gz | sqlite3 chictopia.sqlite3
+	if [ -f "./chictopia.sqlite3" ]
+	then
+		rm chictopia.sql.gz
 	else
-		echo "install tar and run again this command!"
+		echo "install gunzip and sqlite3 and run again this command!"
 		exit 1
 	fi
-else echo "photos database already downloaded!"
-fi
 
-echo "unzipping database.."
-gunzip -c chictopia.sql.gz | sqlite3 chictopia.sqlite3
-if [ -f "./chictopia.sqlite3" ]
-then
-	rm chictopia.sql.gz
+	cd ..
+	cd ..
+	cd ..
+
 else
-	echo "install gunzip and sqlite3 and run again this command!"
-	exit 1
+	echo "Skipping downloading PaperDoll"
 fi
 
-cd ..
-cd ..
-cd ..
 pwd
+
 text="\n
 now downloading modanet annotations\n
 					\t\t\ttaken from here:\n
@@ -68,8 +80,46 @@ git clone https://github.com/eBay/modanet.git
 
 mkdir coco #this will be our dataset final folder
 cd coco
-mkdir images
+
 mkdir annotations
+
+if [ $FAST ]
+then
+	pwd
+	echo "downloading the images folder.. (2 GB)"
+	if [ ! -d "./images" ]
+	then
+		wget -c https://github.com/cad0p/maskrcnn-modanet/releases/download/v0.9/images.zip
+		echo "unzipping.."
+		unzip -q images.zip
+		if [ -d "./images" ]
+		then
+			rm images.zip
+		else
+			echo "could not unzip file. run command again"
+			exit 1
+		fi
+	else
+		echo "images already downloaded!"
+	fi
+
+	cd annotations
+	pwd
+
+	echo "now downloading fixed ModaNet annotations (this can also be done with datasets fix command)"
+	if [ ! -f "./instances_all.json" ]
+	then
+		wget -c https://github.com/cad0p/maskrcnn-modanet/releases/download/v0.9/instances_all.json
+	else
+		echo "fixed ModaNet annotations already downloaded"
+	fi
+	cd ..
+	
+else
+	echo "images will be downloaded afterwards by running datasets arrange command"
+	mkdir images
+fi
+
 
 cd ..
 cd .. #now in main folder
@@ -84,7 +134,13 @@ then
 else echo "default coco snapshot already downloaded"
 fi
 
-# resnet50_modanet.h5
+
+if [ ! -f "./resnet50_modanet.h5" ]
+then
+	echo "downloading the last available trained modanet snapshot"
+	wget -c https://github.com/cad0p/maskrcnn-modanet/releases/download/v0.9/resnet50_modanet.h5
+else echo "default modanet snapshot already downloaded"
+fi
 
 mkdir snapshots
 mkdir processedimages
